pyproject.toml
[project]
name = "rag-notes"
version = "0.1.0"
description = "Local RAG note Q&A assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index>=0.11.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "llama-index-embeddings-openai>=0.2.0",
    "llama-index-llms-openai-like>=0.2.0",
    "gradio>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    "arize-phoenix>=12.28.1",
    "openinference-instrumentation-llama-index>=4.3.9",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "cohere>=4.0.0",
    # Multi-format document support (PDF, Word, Excel, PowerPoint, etc.)
    # Use [all] to include PDF, DOCX, PPTX, XLSX, audio transcription, etc.
    "markitdown[all]>=0.0.1a3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
